+++ src/main/java/net/soulsweaponry/items/sword/WhirligigSawblade.java
@@ -1,100 +1,38 @@
 package net.soulsweaponry.items.sword;
 
-import net.minecraft.enchantment.EnchantmentHelper;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.StatusEffectInstance;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.UseAction;
-import net.minecraft.util.math.Box;
-import net.minecraft.util.math.Vec3d;
-import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
-import net.soulsweaponry.entitydata.BleedData;
 import net.soulsweaponry.items.ModdedSword;
-import net.soulsweaponry.registry.EffectRegistry;
-import net.soulsweaponry.util.TooltipAbilities;
-import net.soulsweaponry.util.WeaponUtil;
-
-import java.util.List;
+import net.soulsweaponry.items.abilities.usagetick.Sawblade;
 
 public class WhirligigSawblade extends ModdedSword {
 
+    private static final Sawblade SAWBLADE = new Sawblade(
+            ConfigConstructor.whirligig_sawblade_ability_range,
+            ConfigConstructor.whirligig_sawblade_ability_expansion,
+            ConfigConstructor.whirligig_sawblade_ability_damage,
+            ConfigConstructor.whirligig_sawblade_ability_bonus_damage_per_level,
+            ConfigConstructor.whirligig_sawblade_ability_enchant_bonus_damage_mod,
+            ConfigConstructor.whirligig_sawblade_knockback,
+            (int) ConfigConstructor.whirligig_sawblade_bleed_added,
+            ConfigConstructor.whirligig_sawblade_bonus_bleed_per_level,
+            (int) ConfigConstructor.whirligig_sawblade_bleed_effect_duration,
+            (int) ConfigConstructor.whirligig_sawblade_bleed_effect_bonus_duration_per_level,
+            (int) ConfigConstructor.whirligig_sawblade_bleed_effect_amp,
+            ConfigConstructor.whirligig_sawblade_bleed_effect_bonus_amp_per_level,
+            (int) ConfigConstructor.whirligig_sawblade_use_time,
+            (int) ConfigConstructor.whirligig_sawblade_bonus_use_time_per_level,
+            (int) ConfigConstructor.whirligig_sawblade_min_cooldown,
+            (int) ConfigConstructor.whirligig_sawblade_cooldown,
+            (int) ConfigConstructor.whirligig_sawblade_reduced_cooldown_per_level
+    );
+
     public WhirligigSawblade(ToolMaterial toolMaterial, Settings settings) {
         super(toolMaterial, (int) ConfigConstructor.whirligig_sawblade_damage, ConfigConstructor.whirligig_sawblade_attack_speed, settings);
-        this.addTooltipAbility(TooltipAbilities.SAWBLADE);
+        this.addAbility(SAWBLADE);
     }
 
-    @Override
-    public UseAction getUseAction(ItemStack stack) {
-        return UseAction.BOW;
-    }
-
-    @Override
-    public int getMaxUseTime(ItemStack stack, LivingEntity user) {
-        return (int) (ConfigConstructor.whirligig_sawblade_use_time + WeaponUtil.getEnchantDamageBonus(stack) * 10);
-    }
-
-    @Override
-    public void usageTick(World world, LivingEntity user, ItemStack stack, int remainingUseTicks) {
-        Vec3d vecBlocksAway = user.getRotationVector().multiply(5).add((user).getPos());
-        Box chunkBox = new Box(user.getX(), user.getY(), user.getZ(), vecBlocksAway.x, vecBlocksAway.y + 1, vecBlocksAway.z);
-        List<Entity> nearbyEntities = world.getOtherEntities(user, chunkBox);
-        if (remainingUseTicks > 0) {
-            for (Entity nearbyEntity : nearbyEntities) {
-                if (nearbyEntity instanceof LivingEntity target && world instanceof ServerWorld serverWorld) {
-                    if (target.damage(world.getDamageSources().mobAttack(user), ConfigConstructor.whirligig_sawblade_ability_damage
-                            + EnchantmentHelper.getDamage(serverWorld, stack, target, world.getDamageSources().mobAttack(user), 0))) {
-                        world.playSound(null, user.getBlockPos(), SoundEvents.ENTITY_PLAYER_ATTACK_SWEEP, SoundCategory.PLAYERS, 1f, 1f);
-                        target.takeKnockback(0.25F, 0, 0);
-                        BleedData.addBleed(target, (int) ConfigConstructor.whirligig_sawblade_bleed_added);
-                        target.addStatusEffect(new StatusEffectInstance(EffectRegistry.BLEED, (int) ConfigConstructor.whirligig_sawblade_bleed_effect_duration, (int) ConfigConstructor.whirligig_sawblade_bleed_effect_amp));
-                    }
-                    world.addParticle(ParticleTypes.SWEEP_ATTACK, true, target.getX(), target.getY() + 1F, target.getZ(), target.getRandom().nextInt(10) - 5, target.getRandom().nextInt(10) - 5, target.getRandom().nextInt(10) - 5);
-                }
-            }
-        } else {
-            user.stopUsingItem();
-            super.usageTick(world, user, stack, remainingUseTicks);
-        }
-    }
-
-    @Override
-    public ItemStack finishUsing(ItemStack stack, World world, LivingEntity user) {
-        this.stop(user, stack);
-        return super.finishUsing(stack, world, user);
-    }
-
-    public void onStoppedUsing(ItemStack stack, World world, LivingEntity user, int remainingUseTicks) {
-        this.stop(user, stack);
-    }
-
-    @Override
-    public boolean canEnchantReduceCooldown(ItemStack stack) {
-        return ConfigConstructor.whirligig_sawblade_enchant_reduces_cooldown;
-    }
-
-    @Override
-    public String[] getReduceCooldownEnchantIds(ItemStack stack) {
-        return ConfigConstructor.whirligig_sawblade_enchant_reduces_cooldown_ids;
-    }
-
-    private void stop(LivingEntity user, ItemStack stack) {
-        if (user instanceof PlayerEntity player) {
-            this.applyItemCooldown(player, this.getCooldown(stack));
-            stack.damage(3, user, WeaponUtil.getActiveHandSlot(player));
-        }
-    }
-
-    private int getCooldown(ItemStack stack) {
-        return (int) Math.max(ConfigConstructor.whirligig_sawblade_min_cooldown, ConfigConstructor.whirligig_sawblade_cooldown - this.getReduceCooldownEnchantLevel(stack) * 10);
-    }
 
     @Override
     public boolean isDisabled(ItemStack stack) {